fix, chyba

+++ CppLibrary/CppLibrary.cpp
@@ -2,5 +2,5 @@
 #include "CppLibrary.h"
 
 float CppConversion(int Red, int Green, int Blue) {
-	return (Red * 0.2627 + Green * 0.6780 + Blue * 0.0593);
+	return (float)(Red * 0.2627 + Green * 0.6780 + Blue * 0.0593);
 }
